Refactored IssueMetadata, Logic, Qualifier and ListPanel.
