+++ src/main/java/ui/listpanel/ListPanel.java
@@ -59,9 +59,16 @@ public class ListPanel extends FilterPanel {
      * @param issue
      * @return true if the issue has changed, false otherwise
      */
+
+    private static boolean qualifierNamesHaveUpdatedQualifier(List<String> filterQualifierNames){
+        // omitting self-updated issues since we only care about updated comments from other users.
+        return (filterQualifierNames.contains(Qualifier.UPDATED) ||
+                filterQualifierNames.contains(Qualifier.UPDATED_OTHERS));
+    }
+
     private boolean issueHasNewComments(TurboIssue issue, boolean hasMetadata) {
         List<String> filterQualifierNames = currentFilterExpression.getQualifierNames();
-        if (hasMetadata && containsUpdatedQualifier(filterQualifierNames)) {
+        if (hasMetadata && qualifierNamesHaveUpdatedQualifier(filterQualifierNames)) {
             return issueNonSelfCommentCounts.containsKey(issue.getId()) &&
                     Math.abs(
                             issueNonSelfCommentCounts.get(issue.getId()) - issue.getMetadata().getNonSelfCommentCount()
@@ -70,12 +77,6 @@ public class ListPanel extends FilterPanel {
             return issueCommentCounts.containsKey(issue.getId()) &&
                     Math.abs(issueCommentCounts.get(issue.getId()) - issue.getCommentCount()) > 0;
         }
-    }
-
-    private static boolean containsUpdatedQualifier(List<String> filterQualifierNames){
-        // omitting self-updated issues since we only care about updated comments from other users.
-        return (filterQualifierNames.contains(Qualifier.UPDATED) ||
-                filterQualifierNames.contains(Qualifier.UPDATED_OTHERS));
     }
 
     /**
